extract AttachmentFrame out of RoomViewInput
